- Added Gaming Platform, Game Genre and Game Filter to Search For Gamers page
- Added ability to filter by Gaming Platform and Game Genre when searching for games

Related work items: #46, #51

+++ PlayTogether/Shared/DTOs/GameSearchDto.cs
@@ -8,8 +8,8 @@ namespace PlayTogether.Shared.DTOs
     {
         public string SearchCriteria { get; set; }
 
-        public List<int> GamingPlatformApiIds { get; set; }
+        public List<long> GamingPlatformApiIds { get; set; }
 
-        public List<int> GameGenreApiIds { get; set; }
+        public List<long> GameGenreApiIds { get; set; }
     }
 }
